Matchings changed

+++ src/main/java/com/volvic20/Volvic20/models/GoogleAPI/payload/ArrivalLocation.java
@@ -1,27 +1,27 @@
 package com.volvic20.Volvic20.models.GoogleAPI.payload;
 
 public class ArrivalLocation {
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
 
-    public ArrivalLocation(float latitude, float longitude) {
+    public ArrivalLocation(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
